- more cleanup and build updates including splitting up more stuff into the x11 package as well as updating the SimpleWorldRenderer for Linux - find OculusSDK cmake updates and linking fixes - initialize and use the Linux Rift SDK - break out the Rift code into its own package - add the sixense directories for Linux and link

+++ iphone_orientation_plugin/iphone_orientation_listener.cpp
@@ -88,7 +88,7 @@ void* setup_socket(void *inputData)
   double buf2[512];
 //  for (i=0; i<NPACK; i++) {
   while(1) {
-	  const int count = recvfrom(s, buf2, BUFLEN, 0, (sockaddr*)&si_other, &slen);
+	  const ssize_t count = recvfrom(s, buf2, BUFLEN, 0, (sockaddr*)&si_other, &slen);
     if (count ==-1)
       diep("recvfrom()");
 //    printf("Received packet from %s:%d\nCount: %d\n\n", inet_ntoa(si_other.sin_addr), ntohs(si_other.sin_port), count);//, buf[0]);
